tetris(6): Allow wall kicks and floor kicks
A kick is a phenomenon that may occur when attempting to rotate a tetromino
fails due to obstacles. If the game can apply some translations to the
position of the rotated tetromino so that it can fit in the board, the
rotation succeeds at the translated position, allowing players to fit
tetrominoes in places where they normally can't fit. Consider the following
case:

  ##    []##  The falling T shape cannot rotate clockwise normally on the
  ##  [][]##  bottom of the field,
  ##    []##
  ####    ##
  ##      ##
  ##########

  ##      ##  because it would collide with the right-hand wall.
  ##      ##
  ##      ##
  ####  []##
  ##  [][][]
  ##########

  ##      ##  However, allowing tetrominoes to kick walls makes it fit in
  ##      ##  the position that is normally unreachable.
  ##      ##
  ####[]  ##
  ##[][][]##
  ##########

There are two popular rotation systems that allow kicks: SRS
(https://harddrop.com/wiki/SRS) and ARS
(https://harddrop.com/wiki/TGM_Rotation). We adopt ARS for the following
reasons:

1. Many experienced Tetris players consider SRS is suitable for multiplayer
   game, because it allows many interesting T-spin tricks, such as TST,
   TSD-Neo, TSD-Iso, and TSD-Fin. However, they (including me) consider
   it's not that great for singleplayer. This is because the rotation rules
   for "S" and "Z" shapes cause a mobility issue under high gravity effect.

2. The current rotation system implemented in tetris(6) is, while not
   exactly the same, very similar to ARS. Tetrominoes spawn non-flat side
   down, "S", "Z", and "I" shapes have only two rotation states as opposed
   to four, and so on. One of the goals of this improvement is to keep the
   game almost unchanged for inexperienced players while providing better
   gameplay experience to veterans, so it's not desirable to change the
   rules drastically.

Note, however, that strictly following the original ARS, where kicking
floor is prohibited, is also not desirable because it relies on
initial-rotation to provide mobility under 20G. Initial-rotation is not
feasible on tty because we cannot get separate events for key-down and
key-up. We therefore modify the rules and allow tetrominoes to kick not
only walls but also floors.