tetris(6): Use a separate timer for "lock delay"
Fall delay is the time duration before moving 1 block down a falling
tetromino. Lock delay is the time before fixing the position of a tetromino
touching the floor. tetris(6) currently uses "fallrate" for both of these
delays. In this change we introduce a separate timer for lock delay. We
define lock delay as 3x longer than fall delay. This makes higher levels
far more playable because you can now shift or rotate tetrominoes that are
already touching the floor for a much longer time.

Also we now only allow tetrominoes to kick the floor at most once. This is
to prevent players from kicking it indefinitely.