genfb(9): New callback function in struct genfb_ops: genfb_dirty_pixels
This is an optional callback invoked when a part of the framebuffer is
updated. GPU drivers that don't support dumb framebuffers can use this to
upload changed parts of a framebuffer allocated in system memory to GPU.

Bump the kernel version to 10.99.9, as it changes the size and the layout
of struct genfb_ops.